Moved SnapshotTrait to Testing namespace.
