+++ src/Testing/SnapshotTrait.php
@@ -2,11 +2,12 @@
 
 declare(strict_types=1);
 
-namespace AlexSkrypnyk\Snapshot;
+namespace AlexSkrypnyk\Snapshot\Testing;
 
 use AlexSkrypnyk\File\File;
 use AlexSkrypnyk\Snapshot\Exception\PatchException;
 use AlexSkrypnyk\Snapshot\Replacer\Replacer;
+use AlexSkrypnyk\Snapshot\Snapshot;
 use PHPUnit\Framework\TestStatus\Error;
 use PHPUnit\Framework\TestStatus\Failure;
