+++ tests/phpunit/UnitTestCase.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace AlexSkrypnyk\Snapshot\Tests;
 
 use AlexSkrypnyk\PhpunitHelpers\UnitTestCase as BaseUnitTestCase;
-use AlexSkrypnyk\Snapshot\SnapshotTrait;
+use AlexSkrypnyk\Snapshot\Testing\SnapshotTrait;
 
 /**
  * Base unit test case for snapshot package tests.
